partial for file exclusions
